Health and damage handling optimization
- There was quite much math done for handling a player's health when
they're damaged. The reason for this was that health variables were
declared as float and a conversion was always needed before handling
everything. This is totally changed now.
- There were useless RPCs sent in order to change money bar value when a
player's health changed. This is useless because we already have a
textdraw for showing HP in numbers. Removed money bar update code.
Removed vending machines as well.
- Checking weapon ranges should be a little faster now because math
(increasing range to avoid some bugs) is done once on server load
instead of every time a player hits someone.
- Little code optimization over OnPlayerTakeDamage callback.
- Fixed a mistake in the previous commit.